f_passport/pyproject.toml
[project]
name = "f-passport"
version = "0.1.0"
description = ""
authors = [
    {name = "Ilya Morozov",email = "morozov1308@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "f_passport", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
